Code for highlevel pickup and drop off system.
Also a temporary pure pursuit controller which I modified from something else. We will want to replace this with whatever Sareum has, or modify the controller to actually stop at pickup and drop off points (since right now this controller always tries to go a set speed).

On top of this can add Amit's collision detection and a lot of pedestrians to not hit.